+++ remyxjs/plugins/math/MathPlugin.js
@@ -303,7 +303,6 @@ export function MathPlugin(options = {}) {
 
     init(eng) {
       engine = eng
-      equationCounter = 0
 
       engine._math = {
         getSymbolPalette,